Add option to enable GNU Install Dirs
This commit adds an option to enable the GNUInstallDirs.cmake script and
use it to define console_bridge's installation directories. When
enabled, libraries, headers, and data are installed to paths defined by
the GNUInstallDirs.cmake script. Otherwise, the hard-coded "lib",
"share", and "include" strings are used, resulting in no change in
functionality after this commit is applied.

The option is diabled by default.

Signed-off-by: Rich Mattes <richmattes@gmail.com>